% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebw.R
\name{ebw}
\alias{ebw}
\title{Alias for \code{\link[=energy_balance]{energy_balance()}}}
\usage{
ebw(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=energy_balance]{energy_balance()}}.}
}
\value{
An object of class \code{"ebw"}.
}
\description{
\code{ebw()} is a thin wrapper around \code{\link[=energy_balance]{energy_balance()}} with identical behavior.
}
