% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ipw_est}
\alias{ipw_est}
\title{Calculation of Hajek estimate of the average treatment effect}
\usage{
ipw_est(y, trt, weights)
}
\arguments{
\item{y}{vector of responses}

\item{trt}{vector indicating treatment assignment of length equal to the length of \code{y}. It should
take the value 1 for treated and any other value to indicate not treated}

\item{weights}{a vector of sample weights of length equal to the length of \code{y}}
}
\value{
an estimate of a causal contrast
}
\description{
Calculates the Hajek weighted estimate of the average treatment effect
}
