% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{print.balance.ebw}
\alias{print.balance.ebw}
\title{Print balance diagnostics}
\usage{
\method{print}{balance.ebw}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{"balance.ebw"} object.}

\item{digits}{number of significant digits.}

\item{...}{unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print balance diagnostics
}
