% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ebw}
\alias{print.ebw}
\title{Print a fitted energy-balancing object}
\usage{
\method{print}{ebw}(x, ...)
}
\arguments{
\item{x}{a fitted \code{"ebw"} object.}

\item{...}{unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted energy-balancing object
}
