% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ebw_object.R
\name{print.energy_balancing_weights}
\alias{print.energy_balancing_weights}
\title{Printing results for estimated energy balancing weights}
\usage{
\method{print}{energy_balancing_weights}(x,
  digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{a fitted object from \code{\link[ebw]{energy_balance}}}

\item{digits}{minimal number of significant digits to print.}

\item{...}{further arguments passed to or from \code{\link[base]{print.default}}.}
}
\description{
Prints results for energy balancing weights
}
\seealso{
\code{\link[ebw]{energy_balance}} for function which produces energy balancing weights
}
