% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_distance.R
\name{weighted_energy_distance}
\alias{weighted_energy_distance}
\title{Calculation of weighted energy distance}
\usage{
weighted_energy_distance(weights = rep(1, NROW(x)), trt, x,
  type = c("two_way", "three_way"), normalize.wts = TRUE)
}
\arguments{
\item{weights}{a vector of sample weights}

\item{trt}{trt vector indicating treatment assignment. Can be a factor or a vector of integers, with each
unique value indicating a different treatment option. When \code{method = 'ATT'}, \code{trt} MUST take
the value \code{1} to indicate treatment.}

\item{x}{matrix of covariates with number of rows equal to the length of \code{weights} and each column is a
covariate}

\item{type}{Which energy distance type to use. if \code{"two_way"}, then the energy distance is the sum of
all of the weighted energy distances between each treatment group and the full sample. If \code{"three_way"},
the energy distance is all of the terms in the \code{"two_way"} distance plus the sum of all weighted energy
distances between every pair of treatment groups.}

\item{normalize.wts}{Defaults to \code{TRUE}, which ensures that the weights within each treatment group
sum to the number of samples in that group. If \code{FALSE}, the weighted energy distance may be ill-defined.}
}
\value{
a scalar value of the requested energy distance
}
\description{
Calculates weighted energy distances between treatment groups and the full sample
}
