% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_distance.R
\name{energy_balance}
\alias{energy_balance}
\title{Construction of energy balancing weights}
\usage{
energy_balance(trt, x, method = c("ATE.3", "ATE", "ATT", "overlap"),
  standardize = TRUE, lambda = 0, verbose = FALSE,
  max.constr = FALSE)
}
\arguments{
\item{trt}{vector indicating treatment assignment. Can be a factor or a vector of integers, with each
unique value indicating a different treatment option. When \code{method = 'ATT'}, \code{trt} MUST take
the value \code{1} to indicate treatment.}

\item{x}{matrix of covariates with number of rows equal to the length of \code{trt} and each column is a
\strong{pre-treatment} covariate to be balanced between treatment groups.}

\item{method}{What method/estimand should be used? \code{"ATE.3"} estimates weights for estimation of the
ATE where the energy distance criterion to be optimized involves all pairwise energy distances. This
method is generally better than \code{"ATE"}, which also estimates the ATE, but only involves energy distances
bbetween each treatment group and the full sample. \code{"ATT"} estimates weights for estimating the ATT}

\item{lambda}{tuning parameter for the penalty on the sum of squares of the weights}

\item{verbose}{should we print out intermediate results of optimization process? \code{TRUE} or \code{FALSE}}

\item{max.constr}{should each weight be constrained to be less than \code{10 * nrow(x) ^ (1/3)}? Defaults to \code{FALSE}.
This option is generally not needed.}

\item{standardized}{logical scalar which results in standardization of the covariates when it takes the value \code{TRUE}
and when it takes the value \code{FALSE}, covariates are not standardized. The default is \code{TRUE}, which is
highly recommended}
}
\value{
An object of class \code{"energy_balancing_weights"} with elements:
\item{weights}{A vector of length \code{nrow(x)} containing the estimates sample weights }
\item{trt}{Treatment vector}
\item{estimand}{The estimand requested}
\item{method}{The method used}
\item{covs}{The covariates that were used for balancing}
\item{energy_dist_unweighted}{The energy distance of the raw data (ie with all weights = 1)}
\item{energy_dist_optimized}{The weighted energy distance using the optimal energy balancing weights}
\item{opt}{The optimization object returned by \code{solvecop()}}
\seealso \code{\link[ebw]{print.energy_balancing_weights}} for printing of fitted energy balancing objects
}
\description{
Constructs energy balancing weights for estimation of average treatment effects
}
\examples{

n <- 100
p <- 5

set.seed(1)

dat <- sim_confounded_data(n.obs = n, n.vars = p, AR.cor = 0.75,
                           propensity.model = "VI", y.model = "D")

x   <- dat$x
y   <- dat$y
trt <- dat$trt

ebal <- energy_balance(trt, x)

# raw energy
ebal$energy_dist_unweighted
# optimal weighted energy
ebal$energy_dist_optimized

# distribution of response:
quantile(y)

# true trt effect:
dat$trt.eff

# naive estimate of trt effect:
ipw_est(y, trt, rep(1, length(trt)))

# estimated trt effect:
ipw_est(y, trt, ebal$weights)

# estimated trt effect with true propensity:
wts_true <- 1 / (trt * dat$prob.trt + (1 - trt) * (1 - dat$prob.trt))
ipw_est(y, trt, wts_true)

}
