% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_confounded_data}
\alias{sim_confounded_data}
\title{Simulation of confounded data}
\usage{
sim_confounded_data(n.obs, n.vars = 10, n.obs.test = 10000,
  AR.cor = 0, propensity.model = c("I", "II", "III", "IV"),
  y.model = c("A", "B", "C", "D", "E"))
}
\arguments{
\item{n.obs}{number of observations}

\item{n.vars}{number of covariates}

\item{n.obs.test}{number of observations for test dataset}

\item{AR.cor}{correlation parameter for AR 1 correlation structure of covariates}

\item{propensity.model}{specification of model for propensity score. Either \code{"I"}, \code{"II"}, or \code{"III"}}

\item{y.model}{specification of model for response. Either \code{"A"}, \code{"B"}, or \code{"C"}}
}
\description{
Simulates confounded data
}
\examples{

n <- 100
p <- 10

set.seed(1)

dat <- sim_confounded_data(n.obs = n, n.vars = p, AR.cor = 0.5,
                           propensity.model = "I", y.model = "I")

x   <- dat$x
y   <- dat$y
trt <- dat$trt

}
