% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_confounded_data}
\alias{sim_confounded_data}
\title{Simulation of confounded data}
\usage{
sim_confounded_data(n.obs, n.vars = 10, n.obs.test = 10000,
  AR.cor = 0, propensity.model = c("I", "II", "III", "IV", "V", "VI",
  "VII"), y.model = c("A", "B", "C", "D", "E", "F"),
  transform.x = FALSE)
}
\arguments{
\item{n.obs}{number of observations}

\item{n.vars}{number of covariates}

\item{n.obs.test}{number of observations for test dataset}

\item{AR.cor}{correlation parameter for AR 1 correlation structure of covariates}

\item{propensity.model}{specification of model for propensity score. Either \code{"I"}, \code{"II"}, or \code{"III"}}

\item{y.model}{specification of model for response. Either \code{"A"}, \code{"B"}, or \code{"C"}}
}
\description{
Simulates confounded data
}
\examples{

n <- 100
p <- 10

set.seed(1)

dat <- sim_confounded_data(n.obs = n, n.vars = p, AR.cor = 0.5,
                           propensity.model = "I", y.model = "I")

x   <- dat$x
y   <- dat$y
trt <- dat$trt

ncols <- 10
Cols <- rbPal(ncols)[as.numeric(cut(dat$prob.trt.test,breaks = ncols))]
plot(dat$x[,1:2], pch=19,cex=1,col=Cols,lwd=2,xlab="",ylab="",
     main = "propensity score vs X1 and X2")
legend("topleft",title="Decile",legend=levels(cut(dat$prob.trt.test,breaks = ncols)),col = rbPal(ncols),pch=19)

plot(dat$x[,2:3], pch=19,cex=1,col=Cols,lwd=2,xlab="",ylab="",
main = "propensity score vs X1 and X2")

plot(dat$x[,3:4], pch=19,cex=1,col=Cols,lwd=2,xlab="",ylab="",
main = "propensity score vs X1 and X2")
legend("bottomright",title="Decile",legend=levels(cut(dat$prob.trt.test,breaks = ncols)),col = rbPal(ncols),pch=19)

plot(dat$x[,4:1], pch=19,cex=1,col=Cols,lwd=2,xlab="",ylab="",
main = "propensity score vs X1 and X2")
legend("bottomright",title="Decile",legend=levels(cut(dat$prob.trt.test,breaks = ncols)),col = rbPal(ncols),pch=19)

}
